Adds GDPR data tables with filtering
Adds data table components for purchases, logins, and transactions to the GDPR section.

Implements filtering, sorting, and pagination for all three tables via new query DTOs.

The date property on the Purchase, Login, and Transaction tables is now sortable.

+++ src-tauri/src/log_parser/types/transaction.rs
@@ -4,7 +4,7 @@ use serde::{Deserialize, Serialize};
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct Transaction {
     pub sku: String,
-    pub price: i64,
+    pub price: f64,
     pub currency: String,
     pub vendor: String,
     pub date: DateTime<Utc>,
